+++ src/usr.sbin/ppp/ipcp.h
@@ -40,13 +40,13 @@ extern struct in_addr nbns_entries[2];
 
 struct ipcpstate {
   struct in_addr his_ipaddr;	/* IP address he is willing to use */
-  u_long his_compproto;
+  u_int32_t his_compproto;
 
   struct in_addr want_ipaddr;	/* IP address I'm willing to use */
-  u_long want_compproto;
+  u_int32_t want_compproto;
 
-  u_long his_reject;		/* Request codes rejected by peer */
-  u_long my_reject;		/* Request codes I have rejected */
+  u_int32_t his_reject;		/* Request codes rejected by peer */
+  u_int32_t my_reject;		/* Request codes I have rejected */
   int heis1172;			/* True if he is speaking rfc1172 */
 };
