+++ src/usr.sbin/ppp/vjcomp.h
@@ -29,4 +29,4 @@
 extern void VjInit(int);
 extern void SendPppFrame(struct mbuf *);
 extern struct mbuf *VjCompInput(struct mbuf *, int);
-extern const char *vj2asc(u_long);
+extern const char *vj2asc(u_int32_t);
